introduced common txs and msgs
