+++ .gitignore
@@ -1,1 +1,4 @@
 node_modules
+
+oracle/.env
+spammer/.env
