+++ common/config/.env.example
@@ -1,7 +1,7 @@
 # Chain
 CHAIN_ID="testing"
 LCD_URL = "http://localhost:1317"
-MNEMONIC = "equip town gesture square tomorrow volume nephew minute witness beef rich gadget actress egg sing secret pole winter alarm law today check violin uncover"
+MNEMONIC = "words in sequential order that unlock your key"
 
 # Price oracle
 MARKET_IDS = "xrp:usd"
