+++ oracle/oracle.js
@@ -1,9 +1,10 @@
-require('dotenv').config();
+require('dotenv').config()
 const cosmosjs = require("@cosmostation/cosmosjs");
 const CoinGecko = require('coingecko-api');
 const cron = require('node-cron');
-import { postTxKava } from './tx.js';
-import { newMsgPostPrice } from './msg.js';
+
+import { postTxKava } from '../common/txs.js';
+import { newMsgPostPrice } from '../common/msg.js';
 import { loadCoinNames } from './utils.js';
 
 // Load chain details, credentials
@@ -37,7 +38,7 @@ var routine = async() => {
         let priceRaw = priceData.data[coinNames[i]].usd
         let price = Number.parseFloat(priceRaw).toPrecision(18).toString();
         // Format msg as JSON
-        let msgPostPrice = newMsgPostPrice("pricefeed/MsgPostPrice", address, marketIDs[i], price)
+        let msgPostPrice = newMsgPostPrice(address, marketIDs[i], price)
         // Send to Kava blockchain
         console.log(coinNames[i], ": posting price", priceRaw)
         postTxKava(kava, chainID, address, ecpairPriv, msgPostPrice)
