+++ spammer/.env
@@ -3,5 +3,9 @@ CHAIN_ID="testing"
 LCD_URL = "http://localhost:1317"
 MNEMONIC = "equip town gesture square tomorrow volume nephew minute witness beef rich gadget actress egg sing secret pole winter alarm law today check violin uncover"
 
+# Price oracle
+MARKET_IDS = "xrp:usd"
+# TODO: Fix async broadcast bug with `MARKET_IDS = "xrp:usd,btc:usd" `
+
 # CDP spammer
 COLLATERAL_DENOM = "xrp"
